Move project to subfolder (Fixes B3zaleel/MangaDownloader#1
