+++ MDXAMLUI/App.xaml
@@ -1,7 +1,7 @@
-﻿<Application x:Class="MangaDownloader.App"
+﻿<Application x:Class="MDXAMLUI.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:MangaDownloader"
+             xmlns:local="clr-namespace:MDXAMLUI"
              StartupUri="MainWindow.xaml">
     <Application.Resources>
         <ResourceDictionary>
