+++ MDXAMLUI/Converters/MangaListItemConverters.cs
@@ -7,7 +7,7 @@ using System.Windows.Data;
 using System.Windows.Media;
 using System.Windows.Media.Imaging;
 
-namespace MangaDownloader.Converters
+namespace MDXAMLUI.Converters
 {
     public class CoverImageToImageSourceConverter : IValueConverter
     {
