+++ MDXAMLUI/IOUtilities/MDBinFormat.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 
-namespace MangaDownloader.IOUtilities
+namespace MDXAMLUI.IOUtilities
 {
     public static class MDBinFormat
     {
