+++ MDXAMLUI/MainWindow.xaml
@@ -1,10 +1,10 @@
-﻿<Window x:Class="MangaDownloader.MainWindow"
+﻿<Window x:Class="MDXAMLUI.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:MangaDownloader"
-        xmlns:conv="clr-namespace:MangaDownloader.Converters"
+        xmlns:local="clr-namespace:MDXAMLUI"
+        xmlns:conv="clr-namespace:MDXAMLUI.Converters"
         xmlns:materialDesign="http://materialdesigninxaml.net/winfx/xaml/themes"
         mc:Ignorable="d"
         TextElement.Foreground="{DynamicResource MaterialDesignBody}"
@@ -51,12 +51,12 @@
 
             <StackPanel Grid.Column="1" HorizontalAlignment="Right" VerticalAlignment="Stretch">
                 <Button Click="ChangeTheme_Click">
-                    <materialDesign:PackIcon 
-                        Grid.Column="0" 
-                        Kind="SunCompass" 
-                        Width="40" 
-                        Height="40" 
-                        VerticalAlignment="Center" 
+                    <materialDesign:PackIcon
+                        Grid.Column="0"
+                        Kind="SunCompass"
+                        Width="40"
+                        Height="40"
+                        VerticalAlignment="Center"
                         HorizontalAlignment="Center"/>
                 </Button>
             </StackPanel>
@@ -77,7 +77,7 @@
                             <RowDefinition Height="60"/>
                         </Grid.RowDefinitions>
                         <ListBox x:Name="mangaListBox" Grid.Row="0" Panel.ZIndex="2" SelectionMode="Single" HorizontalContentAlignment="Stretch"
-                                  ItemsSource="{Binding Mangas, UpdateSourceTrigger=PropertyChanged}" SelectionChanged="MangaListBox_SelectionChanged" 
+                                  ItemsSource="{Binding Mangas, UpdateSourceTrigger=PropertyChanged}" SelectionChanged="MangaListBox_SelectionChanged"
                                  KeyDown="MoveMangaUp_KeyDown">
                             <ListBox.ItemTemplate>
                                 <DataTemplate>
@@ -207,7 +207,7 @@
                         </Grid.RowDefinitions>
 
                         <ListView x:Name="chapterListView" Grid.Row="0" SelectionChanged="ChapterListView_SelectionChanged" SelectionMode="Single" VerticalContentAlignment="Stretch" HorizontalContentAlignment="Stretch"
-                                 ItemsSource="{Binding MangaChapters, BindsDirectlyToSource=True, UpdateSourceTrigger=PropertyChanged}" SelectiveScrollingGrid.SelectiveScrollingOrientation="Vertical" 
+                                 ItemsSource="{Binding MangaChapters, BindsDirectlyToSource=True, UpdateSourceTrigger=PropertyChanged}" SelectiveScrollingGrid.SelectiveScrollingOrientation="Vertical"
                                   KeyDown="ChapterListView_KeyDown">
                             <ListView.ItemTemplate>
                                 <DataTemplate>
@@ -218,7 +218,7 @@
                                         </Grid.ColumnDefinitions>
 
                                         <TextBlock Grid.Column="0" Panel.ZIndex="2" Margin="5 0" Text="{Binding Title}" VerticalAlignment="Center" FontSize="15"/>
-                                        <Border Grid.Column="1" Height="30" 
+                                        <Border Grid.Column="1" Height="30"
                                                 CornerRadius="{Binding Progress, UpdateSourceTrigger=PropertyChanged, Converter={StaticResource ProgressToCornerRadiusConverter}}"
                                                 Margin="{Binding Progress, UpdateSourceTrigger=PropertyChanged, Converter={StaticResource ProgressToMarginConverter}}"
                                                 Background="{Binding Progress, UpdateSourceTrigger=PropertyChanged, Converter={StaticResource ProgressToBackgroundConverter}}">
@@ -242,7 +242,7 @@
                                                     </Path.Style>
                                                 </Path>-->
 
-                                                <TextBlock Text="{Binding Progress, UpdateSourceTrigger=PropertyChanged, Converter={StaticResource ProgressToTextConverter}}" 
+                                                <TextBlock Text="{Binding Progress, UpdateSourceTrigger=PropertyChanged, Converter={StaticResource ProgressToTextConverter}}"
                                                            VerticalAlignment="Center" HorizontalAlignment="Center" Foreground="Lavender" FontSize="12"/>
 
                                                 <Path Margin="-2 -2 0 0" Panel.ZIndex="2" Visibility="{Binding IsComplete, UpdateSourceTrigger=PropertyChanged, Converter={StaticResource CompletionToVisibilityConverter}}" Fill="WhiteSmoke" Stroke="WhiteSmoke" Data="M9 16.2L4.8 12l-1.4 1.4L9 19 21 7l-1.4-1.4L9 16.2z"/>
