+++ MDXAMLUI/Properties/AssemblyInfo.cs
