+++ MDXAMLUI/SearchMangaDialog.xaml
@@ -1,9 +1,9 @@
-﻿<Window x:Class="MangaDownloader.SearchMangaDialog"
+﻿<Window x:Class="MDXAMLUI.SearchMangaDialog"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:MangaDownloader"
+        xmlns:local="clr-namespace:MDXAMLUI"
         xmlns:materialDesign="http://materialdesigninxaml.net/winfx/xaml/themes"
         mc:Ignorable="d"
         TextElement.Foreground="{DynamicResource MaterialDesignBody}"
